Final Commit
Changed many of the variables to store objects in the session storage by going through the form instead of creating a variables and using the getElementsByID("") method. Finished creating the writeSessionValues function to display the cart totals and items. Finished adding in comments to describe my code. Ready for review.